Adding code for images with blue glare
Changes to main code to try to correct for blue glare in some images. (This will be deleted in next push, since it doesn't work well.)

Added functions to address images with blue glare.

Added the FalseColor function to more easily see what the program thinks each layer looks like.

Added a folder of example photos and the output of the layer boundaries...